Add /healthz endpoint
