+++ pkg/server/health.go
@@ -1,0 +1,9 @@
+package server
+
+import "net/http"
+
+func handlerHealth() http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		// For now, we're always healthy, so just let it write 200 OK naturally
+	}
+}
